requirements.txt
beautifulsoup4==4.14.2
bs4==0.0.2
certifi==2025.11.12
charset-normalizer==3.4.4
dnspython==2.8.0
idna==3.11
nvdlib==0.8.3
python-dateutil==2.9.0.post0
python-whois==0.9.6
requests==2.32.5
scapy==2.6.1
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
urllib3==2.5.0
whois==1.20240129.2
